+++ Dockerfile
@@ -7,4 +7,4 @@ WORKDIR "/go/src/github.com/elastic/integrations-registry"
 RUN GO111MODULE=on go mod vendor
 
 # Make sure it's accessible from outside the container
-CMD ["go", "run", "main.go", "--address=0.0.0.0:8080"]
+ENTRYPOINT ["go", "run", ".", "--address=0.0.0.0:8080"]
